scheduler/requirements.txt
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
pytz==2023.3 